refactor: move messages location
