+++ src/studio-home/tabs-section/libraries-tab/MigrateLegacyLibrariesAlert.tsx
@@ -2,7 +2,7 @@ import { Alert, Button } from '@openedx/paragon';
 import { Warning } from '@openedx/paragon/icons';
 import { FormattedMessage } from '@edx/frontend-platform/i18n';
 
-import messages from './messages';
+import messages from '../messages';
 
 export const MigrateLegacyLibrariesAlert = () => (
   <Alert variant="warning" icon={Warning}>
